file rename, move cam settins to func in config
